egrfwetfqưẻtwq

+++ bienvakieudulieu/bai9.cpp
@@ -2,8 +2,8 @@
 using namespace std;
 
 int main (){
-    long a = 12984019;
-    long b = 12948120;
+    int a = 12984019;
+    int b = 12948120;
     long p = a*b;
     cout << p;
